Refactor authentication logic to improve password comparison and streamline user session handling

+++ auth.ts
@@ -1,5 +1,155 @@
+// import { PrismaAdapter } from "@auth/prisma-adapter";
+// import { compareSync } from "bcrypt-ts-edge";
+// import NextAuth, { NextAuthConfig } from "next-auth";
+// import CredentialsProvider from "next-auth/providers/credentials";
+// import { cookies } from "next/headers";
+// import { NextResponse } from "next/server";
+// import { prisma } from "./db/prisma";
+
+// export const config = {
+//   pages: {
+//     signIn: "/sign-in",
+//     error: "/error",
+//   },
+//   session: {
+//     strategy: "jwt",
+//     maxAge: 30 * 24 * 60 * 60,
+//   },
+//   adapter: PrismaAdapter(prisma),
+//   providers: [
+//     CredentialsProvider({
+//       credentials: {
+//         email: { type: "email" },
+//         password: { type: "password" },
+//       },
+//       async authorize(credentials) {
+//         if (credentials == null) return null;
+//         //Find user in database
+//         const user = await prisma.user.findFirst({
+//           where: {
+//             email: credentials.email as string,
+//           },
+//         });
+//         //check if user exists & if password matches
+//         if (user && user.password) {
+//           const isMatch = compareSync(
+//             credentials.password as string,
+//             user.password
+//           );
+//           //If password correct return user
+//           if (isMatch) {
+//             return {
+//               id: user.id,
+//               name: user.name,
+//               email: user.email,
+//               role: user.role,
+//             };
+//           }
+//         }
+//         //if user does not exists or password does not match return null
+//         return null;
+//       },
+//     }),
+//   ],
+//   callbacks: {
+//     async session({ session, user, trigger, token }: any) {
+//       //set the userId from the token
+//       session.user.id = token.sub;
+//       session.user.role = token.role;
+//       session.user.name = token.name;
+
+//       //if there as an update, set the user name
+//       if (trigger === "update") {
+//         session.user.name = user.name;
+//       }
+//       return session;
+//     },
+
+//     async jwt({ token, user, trigger, session }: any) {
+//       //Assign user fields to token
+//       if (user) {
+//         token.id = user.id;
+//         token.role = user.role;
+//         //if user have no name, use the email
+//         if (user.name === "NO_NAME") {
+//           token.name = user.email!.split("@")[0];
+//           //Update database to reflect the token name
+//           await prisma.user.update({
+//             where: { id: user.id },
+//             data: { name: token.name },
+//           });
+//         }
+//         if (trigger === "signIn" || trigger === "signUp") {
+//           const cookiesObject = await cookies();
+//           const sessionCartId = cookiesObject.get("sessionCartId")?.value;
+//           if (sessionCartId) {
+//             const sessionCart = await prisma.cart.findFirst({
+//               where: { sessionCartId },
+//             });
+//             if (sessionCart) {
+//               //Delete current user cart
+//               await prisma.cart.deleteMany({
+//                 where: { userId: user.id },
+//               });
+//               //Assign new cart
+//               await prisma.cart.update({
+//                 where: { id: sessionCart.id },
+//                 data: { userId: user.id },
+//               });
+//             }
+//           }
+//         }
+//       }
+//       //Handle session update
+//       if (session?.user.name && trigger === "update") {
+//         token.name = session.user.name;
+//       }
+//       return token;
+//     },
+
+//     authorized({ request, auth }: any) {
+//       //Array of regex patterns of paths we want to protect
+//       const protectedPaths = [
+//         /\/shipping-address/,
+//         /\/payment-methods/,
+//         /\/place-order/,
+//         /\/profile/,
+//         /\/user\/(.*)/,
+//         /\/admin/,
+//       ];
+//       //Get pathname from request URL object
+//       const { pathname } = request.nextUrl;
+//       //Check if user is not authorized and accessing protected paths
+//       if (!auth && protectedPaths.some((path) => path.test(pathname))) {
+//         return false;
+//       }
+
+//       //Check for session cart cookie
+//       if (!request.cookies.get("sessionCartId")) {
+//         //Generate new session cart id cookie
+//         const sessionCartId = crypto.randomUUID();
+//         //clone the req headers
+//         const newRequestHeaders = new Headers(request.headers);
+//         //create new response and add the new headers
+//         const response = NextResponse.next({
+//           request: {
+//             headers: newRequestHeaders,
+//           },
+//         });
+//         //Set newly generated sessionCartId in the response cookies
+//         response.cookies.set("sessionCartId", sessionCartId);
+//         return response;
+//       } else {
+//         return true;
+//       }
+//     },
+//   },
+// } satisfies NextAuthConfig;
+
+// export const { handlers, auth, signIn, signOut } = NextAuth(config);
+
+import { comparePasswords } from "@/lib/bcryptUtil";
 import { PrismaAdapter } from "@auth/prisma-adapter";
-import { compareSync } from "bcrypt-ts-edge";
 import NextAuth, { NextAuthConfig } from "next-auth";
 import CredentialsProvider from "next-auth/providers/credentials";
 import { cookies } from "next/headers";
@@ -23,20 +173,18 @@ export const config = {
         password: { type: "password" },
       },
       async authorize(credentials) {
-        if (credentials == null) return null;
-        //Find user in database
+        if (!credentials) return null;
+
         const user = await prisma.user.findFirst({
-          where: {
-            email: credentials.email as string,
-          },
+          where: { email: credentials.email as string },
         });
-        //check if user exists & if password matches
-        if (user && user.password) {
-          const isMatch = compareSync(
+
+        if (user?.password) {
+          const isMatch = await comparePasswords(
             credentials.password as string,
             user.password
           );
-          //If password correct return user
+
           if (isMatch) {
             return {
               id: user.id,
@@ -46,19 +194,16 @@ export const config = {
             };
           }
         }
-        //if user does not exists or password does not match return null
         return null;
       },
     }),
   ],
   callbacks: {
     async session({ session, user, trigger, token }: any) {
-      //set the userId from the token
       session.user.id = token.sub;
       session.user.role = token.role;
       session.user.name = token.name;
 
-      //if there as an update, set the user name
       if (trigger === "update") {
         session.user.name = user.name;
       }
@@ -66,32 +211,28 @@ export const config = {
     },
 
     async jwt({ token, user, trigger, session }: any) {
-      //Assign user fields to token
       if (user) {
         token.id = user.id;
         token.role = user.role;
-        //if user have no name, use the email
+
         if (user.name === "NO_NAME") {
           token.name = user.email!.split("@")[0];
-          //Update database to reflect the token name
           await prisma.user.update({
             where: { id: user.id },
             data: { name: token.name },
           });
         }
+
         if (trigger === "signIn" || trigger === "signUp") {
           const cookiesObject = await cookies();
           const sessionCartId = cookiesObject.get("sessionCartId")?.value;
+
           if (sessionCartId) {
             const sessionCart = await prisma.cart.findFirst({
               where: { sessionCartId },
             });
             if (sessionCart) {
-              //Delete current user cart
-              await prisma.cart.deleteMany({
-                where: { userId: user.id },
-              });
-              //Assign new cart
+              await prisma.cart.deleteMany({ where: { userId: user.id } });
               await prisma.cart.update({
                 where: { id: sessionCart.id },
                 data: { userId: user.id },
@@ -100,7 +241,7 @@ export const config = {
           }
         }
       }
-      //Handle session update
+
       if (session?.user.name && trigger === "update") {
         token.name = session.user.name;
       }
@@ -108,7 +249,6 @@ export const config = {
     },
 
     authorized({ request, auth }: any) {
-      //Array of regex patterns of paths we want to protect
       const protectedPaths = [
         /\/shipping-address/,
         /\/payment-methods/,
@@ -117,31 +257,22 @@ export const config = {
         /\/user\/(.*)/,
         /\/admin/,
       ];
-      //Get pathname from request URL object
       const { pathname } = request.nextUrl;
-      //Check if user is not authorized and accessing protected paths
+
       if (!auth && protectedPaths.some((path) => path.test(pathname))) {
         return false;
       }
 
-      //Check for session cart cookie
       if (!request.cookies.get("sessionCartId")) {
-        //Generate new session cart id cookie
         const sessionCartId = crypto.randomUUID();
-        //clone the req headers
         const newRequestHeaders = new Headers(request.headers);
-        //create new response and add the new headers
         const response = NextResponse.next({
-          request: {
-            headers: newRequestHeaders,
-          },
+          request: { headers: newRequestHeaders },
         });
-        //Set newly generated sessionCartId in the response cookies
         response.cookies.set("sessionCartId", sessionCartId);
         return response;
-      } else {
-        return true;
       }
+      return true;
     },
   },
 } satisfies NextAuthConfig;
